implementing Task 4 in v1/places, v1/reviews, v1/users

+++ part3/hbnb/app/api/v1/reviews.py
@@ -1,5 +1,6 @@
 from flask_restx import Namespace, Resource, fields
 from app.services import facade
+from flask_jwt_extended import jwt_required, get_jwt_identity, get_jwt
 
 api = Namespace('reviews', description='Review operations')
 
@@ -7,7 +8,7 @@ api = Namespace('reviews', description='Review operations')
 review_model = api.model('Review', {
     'text': fields.String(required=True, description='Text of the review'),
     'rating': fields.Integer(required=True, description='Rating of the place (1-5)'),
-    'user_id': fields.String(required=True, description='ID of the user'),
+    'user_id': fields.String(description='ID of the user'),
     'place_id': fields.String(required=True, description='ID of the place')
 })
 
@@ -16,17 +17,28 @@ class ReviewList(Resource):
     @api.expect(review_model)
     @api.response(201, 'Review successfully created')
     @api.response(400, 'Invalid input data')
+    @jwt_required()
     def post(self):
         """Register a new review"""
         review_data = api.payload
+
+        user = get_jwt_identity()
+        review_data["user_id"] = user
+
+        claims = get_jwt()
+        is_admin = claims.get("is_admin", False)
+
         place = facade.get_place(review_data['place_id'])
         if not place:
             return {'error': 'Place not found'}, 400
-        user = facade.get_user(review_data['user_id'])
-        if not user:
-            return {'error': 'User not found'}, 400
-        if place.owner.id == user.id:
-            return {'error': 'User cannot review their own place'}, 400
+
+        if not is_admin and place.owner.id == user:
+            return {'error': 'You cannot review your own place'}, 400
+        
+        for review in place.reviews:
+            if not is_admin and review.user_id == user:
+                return {'error': 'You have already reviewed this place.'}, 400
+
         try:
             new_review = facade.create_review(review_data)
             return new_review.to_dict(), 201
@@ -53,13 +65,20 @@ class ReviewResource(Resource):
     @api.response(200, 'Review updated successfully')
     @api.response(404, 'Review not found')
     @api.response(400, 'Invalid input data')
+    @jwt_required()
     def put(self, review_id):
         """Update a review's information"""
         review_data = api.payload
         review = facade.get_review(review_id)
         if not review:
             return {'error': 'Review not found'}, 404
-        
+
+        user = get_jwt_identity()
+        claims = get_jwt()
+        is_admin = claims.get("is_admin", False)
+        if not is_admin and review.user_id != user:
+            return {'error': 'Unauthorized action'}, 403
+
         try:
             facade.update_review(review_id, review_data)
             return {'message': 'Review updated successfully'}, 200
@@ -68,12 +87,21 @@ class ReviewResource(Resource):
 
     @api.response(200, 'Review deleted successfully')
     @api.response(404, 'Review not found')
+    @jwt_required()
     def delete(self, review_id):
         """Delete a review"""
         review = facade.get_review(review_id)
         if not review:
             return {'error': 'Review not found'}, 404
-        
+
+        user = get_jwt_identity()
+        claims = get_jwt()
+        is_admin = claims.get("is_admin", False)
+
+        if not is_admin and review.user_id != user:
+            return {'error': 'Unauthorized action'}, 403
+
+
         try:
             facade.delete_review(review_id)
             return {'message': 'Review deleted successfully'}, 200
